Improved kill feeds in terms of performance and query minimization

+++ ThunderED/Json/ZKill/Killmail.cs
@@ -6,11 +6,11 @@ namespace ThunderED.Json.ZKill
     {
         public class Killmail
         {
-            public int killmail_id { get; set; }
+            public long killmail_id { get; set; }
             public DateTime killmail_time { get; set; }
             public Victim victim { get; set; }
             public Attacker[] attackers { get; set; }
-            public int solar_system_id { get; set; }
+            public long solar_system_id { get; set; }
             public Zkb zkb;
         }
     }
